GENAI_FSD/auto_invest_AI/backend/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
